feat: add delete user images when delete user

+++ lib/services/user_service.dart
@@ -1,8 +1,10 @@
 import 'package:geradordeimagem_back_dart/repositories/user_repository.dart';
+import 'package:geradordeimagem_back_dart/services/image_service.dart';
 
 class UserService {
 
   final repo = UserRepository();
+  final imageService = ImageService();
 
   signUp(userData) async {
     final hashedPassword = repo.hashPassword(userData['password']);
@@ -18,6 +20,17 @@ class UserService {
   }
 
   removeUser(email) async {
+    final user = await repo.getUserByEmail(email);
+    
+    if (user == null) {
+      throw Exception('User not found');
+    }
+
+    final userId = user['_id'];
+    
+    final imagesDeletion = await imageService.deleteUserImages(userId);
+    print('Deleted ${imagesDeletion['deletedCount']} images for user $email');
+
     return await repo.removeUser(email);
   }
 
@@ -27,5 +40,13 @@ class UserService {
 
   getUserByEmail(email) async {
     return await repo.getUserByEmail(email);
+  }
+
+  updateUser(String email, Map<String, dynamic> updateData, {bool isAdmin = false}) async {
+    if (!isAdmin && updateData.containsKey('admin')) {
+      updateData.remove('admin');
+    }
+
+    return await repo.updateUser(email, updateData);
   }
 }
